+++ .gitignore
@@ -1,0 +1,2 @@
+# Output that does not need committing
+cluster.txt
