+++ README.md
@@ -4,9 +4,18 @@ Readme
 Instructions
 ---------------------
 
-Timepix folder contains the original code while Timepixeval contains the working code. After creating a code that works in Timepixeval, we save the changes to Timepix. Therefore, Timepixeval would contain the most recently updated codes. 
-
 ### How to use
 
-> 1) Run createdatabase.py (this step is optional as I have done it and there is already a database.txt but it is up to you.)
-> 2) Run threshold_and_matching.py (you should see a graph of some sort showing the amount of matches)
+* To create the database: 
+
+```bash
+$ python createdatabase.py
+```
+
+* To apply the threshold and matching procedure:
+
+```bash
+$ python threshold_and_matching.py
+```
+
+You should see a graph of some sort showing the amount of matches.
